vokaturi_service/requirements.txt
Flask==1.1.1
scipy==1.3.2
gunicorn==19.9.0
numpy==1.17.3
google-cloud-storage==1.19.1